Update the receive email hook to accept email from anonymous watchers
Since the watchers are locked (and we want to keep them that way) the
original redmine code rejected them as senders. This patch will make
sure they are temporarily activated for the duration of the receive
code.

It would probably be better if the name of the "Anonymous Watchers"
group was a constant in the code somewhere, not sure how to do that,
but this seems to work for the moment.