Make it possible to add existing anonymous wathcers
We create a new user for each anonymous wathcers (once emails are sent).
This caused the previous code to try to add inactive low privilige
accounts to the list of anonymous wathcers instead of just adding
another row to the anonymous_watchers table. That didn't work since that
user didn't have permission to watch in the project.

This patch makes sure that only users that are not anonymous wathcers
are added as normal wathcers and that anonymous wathcers gets added by
email.